Provide the ability to opt out of DS.Adapter prototype extension
Updates the addon such that consumers have the ability to opt out
of deafultUUID's via prototype extension and opt in to using an
adapter mixin at a granularity of their choosing (i.e. globally via
an ApplicationSerializer or on an Adapter by Adapter basis). In the
process of adding the new feature I also moved the UUID regex tester
into a reusable helper. Lastly, it's worth noting that I didn't
remove the adapter reopen completely as doing so would compromise
the functionality of the Configuration.js module. Instead I kept
the spirit of the original implementation and updated the reopen
to conditionally delegate the generatedIdForRecord call to the
same function that the adapter mixin both uses and exports for a
use case such as this.